requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
dash
plotly
pandas
psycopg2-binary
